requirements.txt
pandas==2.3.0
numpy==2.2.6
sentence-transformers==4.1.0
chromadb==1.0.12
google-generativeai==0.8.5
tqdm==4.67.1
fastapi==0.115.9
uvicorn==0.34.3
pyngrok==7.2.11
nest-asyncio==1.6.0
kagglehub==0.3.12
scikit-learn==1.7.0
scipy==1.15.3
python-dotenv==1.1.0
requests==2.32.4
joblib==1.5.1
transformers==4.52.4
torch==2.7.1
